Move a bunch of stuff around.
