+++ fresnel/test/Test.hs
@@ -11,7 +11,7 @@ module Main
 import           Control.Carrier.Reader
 import           Control.Carrier.State.Church
 import           Control.Exception (SomeException)
-import           Control.Monad (guard, join, when)
+import           Control.Monad (guard, when)
 import           Control.Monad.IO.Class
 import           Data.Bool (bool)
 import           Data.Foldable (for_, toList)
@@ -22,8 +22,6 @@ import           Data.Maybe (fromMaybe)
 import           Data.Monoid (Ap(..))
 import           Data.Ord (comparing)
 import qualified Fold.Test
-import           Fresnel.Maybe (_Just)
-import           Fresnel.Optional (is)
 import           Fresnel.Tropical
 import           GHC.Exception.Type (Exception(displayException))
 import qualified Getter.Test
@@ -37,6 +35,7 @@ import           System.Environment (getArgs)
 import           System.IO
 import           Test.Group as Group
 import           Test.Options
+import           Test.Print
 import           Test.QuickCheck (Args(..), isSuccess, quickCheckWithResult)
 import qualified Test.QuickCheck as QC
 import qualified Tropical.Test
@@ -232,92 +231,3 @@ runTally t = sepBy_ (putS ", " ) (map success (tally "✓" "success" "successes"
 
 sepBy_ :: Applicative m => m () -> [m ()] -> m ()
 sepBy_ sep = getAp . foldMap Ap . intersperse sep
-
-
-hasSuccesses :: Tally -> Bool
-hasSuccesses = (/= 0) . successes
-
-hasFailures :: Tally -> Bool
-hasFailures = (/= 0) . failures
-
-unit :: Status -> Tally
-unit = stat (Tally 1 0) (Tally 0 1)
-
-data Tally = Tally { successes :: Int, failures :: Int }
-
-instance Semigroup Tally where
-  Tally s1 f1 <> Tally s2 f2 = Tally (s1 + s2) (f1 + f2)
-
-instance Monoid Tally where
-  mempty = Tally 0 0
-
-
-withSGR :: (Has (Reader Handle) sig m, MonadIO m) => [SGR] -> m a -> m a
-withSGR sgr m = withHandle $ \ h -> liftIO (hSetSGR h sgr) *> m <* liftIO (hSetSGR h [])
-
-colour :: (Has (Reader Handle) sig m, MonadIO m) => ColorIntensity -> Color -> m a -> m a
-colour i c = withSGR [SetColor Foreground i c]
-
-success, failure, failure' :: (Has (Reader Handle) sig m, MonadIO m) => m a -> m a
-
-success = colour Vivid Green
-failure = colour Vivid Red
-failure' = colour Dull Red
-
-
-sparks :: String
-sparks = " ▁▂▃▄▅▆▇█"
-
-sparkify :: Real a => [a] -> String
-sparkify bins = sparkifyRelativeTo sparks (maximum bins) bins
-
-sparkifyRelativeTo :: Real a => String -> a -> [a] -> String
-sparkifyRelativeTo sparks max = fmap spark
-  where
-  spark n = sparks !! round (realToFrac n / realToFrac max * realToFrac (length sparks - 1) :: Double)
-
-
-stat :: a -> a -> Status -> a
-stat pass fail = \case{ Pass -> pass ; Fail -> fail }
-
-data Status = Pass | Fail
-  deriving (Eq)
-
-
-topIndent :: (Has (Reader Handle) sig m, Has (State Tally) sig m, MonadIO m) => String -> m ()
-topIndent m = gets hasFailures >>= failure' . putS . bool space m
-
-withHandle :: Has (Reader Handle) sig m => (Handle -> m a) -> m a
-withHandle = join . asks
-
-
-line :: (Has (Reader Handle) sig m, Has (State Tally) sig m, MonadIO m) => m ()
-line = topIndent vline *> putS vline
-
-section :: (Has (Reader Handle) sig m, Has (Reader Width) sig m, Has (State Tally) sig m, MonadIO m) => Maybe Status -> m a -> m a
-section s m = do
-  fullWidth <- asks ((+ (3 + length "Success")) . width)
-  let rule corner = indent *> maybe id (stat success failure) s (putS (corner : h : replicate fullWidth h)) *> nl
-  rule '╭' *> m <* rule '╰'
-  where
-  indent = topIndent vline *> when (is _Just s) (putS vline)
-  h = '─'
-
-
-space, bullet, heading1, headingN, arrow, vline, hline, gtally, end :: String
-space    = "  "
-bullet   = "☙ "
-heading1 = "╭─"
-headingN = "├─"
-arrow    = "▶ "
-vline    = "│ "
-hline    = "──"
-gtally   = "┤ "
-end      = "╰─┤ "
-
-
-nl :: (Has (Reader Handle) sig m, MonadIO m) => m ()
-nl = withHandle (liftIO . (`hPutStrLn` ""))
-
-putS :: (Has (Reader Handle) sig m, MonadIO m) => String -> m ()
-putS s = withHandle (liftIO . (`hPutStr` s))
